add a test for the oath case in generator
