+++ lib/auth/github.js
@@ -39,7 +39,7 @@ GithubAuth.prototype.authBasic = function(user, pass){
 };
 
 
-GithubAuth.prototype.getAuthToken = function(optionalTwoFactor){
+GithubAuth.prototype.fetchAndStoreAuthToken = function(optionalTwoFactor){
   var self = this;
   return new BluePromise(function(s, f){
     var headers = {};    
@@ -71,6 +71,26 @@ GithubAuth.prototype.getAuthToken = function(optionalTwoFactor){
     });  
   });
 };
+
+
+GithubAuth.prototype.readAuthToken = function(){
+  var dotfile = Dotfile('repomanrc');
+  return new BluePromise(function(s, f){   
+      dotfile.exists(function (yesno) {
+        if(yesno){
+          dotfile.read(function (err, disk) {
+            if(disk.githubAuthToken){           
+              s(disk.githubAuthToken);
+            } else {
+              f(new Error('no auth token in dotfile'));
+            }
+          });
+        } else {
+          f(new Error('no dotfile'));
+        }
+      });
+    });
+}
 
 
 module.exports = GithubAuth;
