+++ lib/cli.js
@@ -48,10 +48,10 @@ function _signin(args) {
         authOpts.password = result['Enter your Github password'];
         var ghAuth = new GithubAuth();
         ghAuth.authBasic(authOpts.username, authOpts.password);
-        ghAuth.getAuthToken(authOpts.twoFactor).catch(function(e){
+        ghAuth.fetchAndStoreAuthToken(authOpts.twoFactor).catch(function(e){
           if(e && e.code === 401 && e.message && e.message.indexOf('two-factor' ) !== -1){
             prompt.get(['Enter your Github two factor auth code'], function (err, result) {
-              ghAuth.getAuthToken(result['Enter your Github two factor auth code']);
+              ghAuth.fetchAndStoreAuthToken(result['Enter your Github two factor auth code']);
             });  
           }          
         });
