+++ lib/generator/github.js
@@ -3,7 +3,7 @@ var bag     = require('bagofrequest');
 var github  = require('github');
 var url     = require('url');
 var BluePromise = require('bluebird');
-var Dotfile = require('dotfile');
+var GithubAuth = require('../auth/github');
 
 /**
  * class GitHub
@@ -40,27 +40,14 @@ function GitHub(user, pass, ready) {
     ready();
     return;
   }
-  var dotfile = Dotfile('repomanrc');
-  var oauthPromise = new BluePromise(function(s, f){
-    dotfile.exists(function (yesno) {
-      if(yesno){
-        dotfile.read(function (err, disk) {
-          if(disk.githubAuthToken){
-            self.gh.authenticate({
-              type : 'oauth',
-              token : disk.githubAuthToken
-            });
-            s();
-          } else {
-            f(new Error('no auth token in dotfile'));
-          }
-        });
-      } else {
-        f(new Error('no dotfile'));
-      }
-    });
-  });
-  oauthPromise.then(ready, function error(){
+ 
+  (new GithubAuth()).readAuthToken().then(function success(token){
+     self.gh.authenticate({
+        type : 'oauth',
+        token : token
+      });
+      ready();
+  }, function error(){
     ready();
   });
 }
